+++ CHANGELOG.md
@@ -20,6 +20,7 @@ to [Semantic Versioning][semver].
 - The MIT license was assigned upon creation of this project and contained code.
 - This CHANGELOG was started with initial entries as the repo is being created.
 - Some notes were left for maintaining repositiories and related contributions.
+- Confirmation in setup for this plugin was gathered with a nice new test suite.
 
 <!-- a collection of links -->